Autoconversion test refactor and clean-up.
