+++ components/scream/src/physics/p3/tests/CMakeLists.txt
@@ -13,6 +13,7 @@ set(P3_TESTS_SRCS
   p3_ice_collection_unit_tests.cpp
   p3_rain_sed_unit_tests.cpp
   p3_dsd2_unit_tests.cpp
+  p3_autoconversion_tests.cpp
 )
 CreateUnitTest(p3_tests "${P3_TESTS_SRCS}" "${NEED_LIBS}" THREADS 1 ${SCREAM_TEST_MAX_THREADS} ${SCREAM_TEST_THREAD_INC} DEP p3_tests_ut_np1_omp1)
